Move all user picture logic to sas
